split up NestedDataColumnSerializer into separate files
